# Optimize Combine for all nil scenarios