--- a/benchmarks_test.go
+++ b/benchmarks_test.go
@@ -76,14 +76,14 @@ func BenchmarkCombine(b *testing.B) {
 		}
 	})
 
-	b.Run("inline 3", func(b *testing.B) {
+	b.Run("inline 3 no error", func(b *testing.B) {
 		var x, y, z error
 		for i := 0; i < b.N; i++ {
 			Combine(x, y, z)
 		}
 	})
 
-	b.Run("inline 3 with error", func(b *testing.B) {
+	b.Run("inline 3 one error", func(b *testing.B) {
 		var x, y, z error
 		z = fmt.Errorf("failed")
 		for i := 0; i < b.N; i++ {
@@ -91,10 +91,36 @@ func BenchmarkCombine(b *testing.B) {
 		}
 	})
 
-	b.Run("slice", func(b *testing.B) {
+	b.Run("inline 3 multiple errors", func(b *testing.B) {
 		var x, y, z error
+		z = fmt.Errorf("failed3")
+		y = fmt.Errorf("failed2")
+		x = fmt.Errorf("failed")
 		for i := 0; i < b.N; i++ {
-			errs := []error{x, y, z}
+			Combine(x, y, z)
+		}
+	})
+
+	b.Run("slice 100 no errors", func(b *testing.B) {
+		for i := 0; i < b.N; i++ {
+			errs := make([]error, 100)
+			Combine(errs...)
+		}
+	})
+
+	b.Run("slice 100 one error", func(b *testing.B) {
+		for i := 0; i < b.N; i++ {
+			errs := make([]error, 100)
+			errs[len(errs)-1] = fmt.Errorf("failed")
+			Combine(errs...)
+		}
+	})
+
+	b.Run("slice 100 multi error", func(b *testing.B) {
+		for i := 0; i < b.N; i++ {
+			errs := make([]error, 100)
+			errs[0] = fmt.Errorf("failed1")
+			errs[len(errs)-1] = fmt.Errorf("failed2")
 			Combine(errs...)
 		}
 	})