During the work on https://github.com/yarpc/yarpc-go/pull/2126 I found that `multierr.Combine` always allocates the error slice on the heap because of the escape analysis.

Assuming that most cases where `multierr.Combine` is called all arguments are `nil` it make sense to optimize it with that in mind: 

Benchmark results for this optimization:
```
name                              old time/op    new time/op    delta
Combine/inline_1                    29.3ns ±15%     2.0ns ± 7%   -93.28%  (p=0.000 n=10+9)
Combine/inline_2                    40.0ns ± 6%     3.4ns ±11%   -91.55%  (p=0.000 n=10+10)
Combine/inline_3_no_error           41.9ns ± 2%     4.8ns ±15%   -88.41%  (p=0.000 n=8+10)
Combine/inline_3_one_error          41.6ns ± 3%     5.3ns ± 5%   -87.18%  (p=0.000 n=9+10)
Combine/inline_3_multiple_errors    81.0ns ± 9%   115.8ns ±16%   +42.96%  (p=0.000 n=10+10)
Combine/slice_100_no_errors          432ns ±12%      99ns ± 8%   -77.20%  (p=0.000 n=10+10)
Combine/slice_100_one_error          555ns ±12%     182ns ± 6%   -67.15%  (p=0.000 n=10+9)
Combine/slice_100_multi_error        832ns ± 6%     919ns ± 7%   +10.38%  (p=0.000 n=10+10)

name                              old alloc/op   new alloc/op   delta
Combine/inline_1                     16.0B ± 0%      0.0B       -100.00%  (p=0.000 n=10+10)
Combine/inline_2                     32.0B ± 0%      0.0B       -100.00%  (p=0.000 n=10+10)
Combine/inline_3_no_error            48.0B ± 0%      0.0B       -100.00%  (p=0.000 n=10+10)
Combine/inline_3_one_error           48.0B ± 0%      0.0B       -100.00%  (p=0.000 n=10+10)
Combine/inline_3_multiple_errors     80.0B ± 0%     80.0B ± 0%      ~     (all equal)
Combine/slice_100_no_errors         1.79kB ± 0%    0.00kB       -100.00%  (p=0.000 n=10+10)
Combine/slice_100_one_error         1.82kB ± 0%    0.02kB ± 0%   -98.68%  (p=0.000 n=10+10)
Combine/slice_100_multi_error       1.90kB ± 0%    1.90kB ± 0%      ~     (all equal)

name                              old allocs/op  new allocs/op  delta
Combine/inline_1                      1.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
Combine/inline_2                      1.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
Combine/inline_3_no_error             1.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
Combine/inline_3_one_error            1.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
Combine/inline_3_multiple_errors      2.00 ± 0%      2.00 ± 0%      ~     (all equal)
Combine/slice_100_no_errors           1.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
Combine/slice_100_one_error           3.00 ± 0%      2.00 ± 0%   -33.33%  (p=0.000 n=10+10)
Combine/slice_100_multi_error         7.00 ± 0%      7.00 ± 0%      ~     (all equal)
```