@abhinav 2022-02-28
Nice one @jstroem, but I don't think it needs to be this complicated.
The issue is that returning the input `errors` inside the multierr puts it on the heap.
So we can simplify this change significantly for the same gains by making a copy of errors for that case only.
I tried making the change and here are the results with the updated benchmarks:


```
name \ time/op                      master.txt   original_pr.txt  simplified_pr.txt
Combine/inline_1-8                  17.7ns ± 0%       2.0ns ± 0%         3.0ns ± 0%
Combine/inline_2-8                  21.0ns ± 0%       3.6ns ± 1%         4.1ns ± 1%
Combine/inline_3_no_error-8         24.4ns ± 0%       4.4ns ± 1%         4.7ns ± 1%
Combine/inline_3_one_error-8        24.8ns ± 0%       4.1ns ± 0%         5.1ns ± 1%
Combine/inline_3_multiple_errors-8  44.3ns ± 0%      60.8ns ± 0%        55.1ns ± 1%
Combine/slice_100_no_errors-8       72.9ns ± 0%      71.3ns ± 0%        72.9ns ± 0%
Combine/slice_100_one_error-8       74.5ns ± 0%      73.7ns ± 0%        74.4ns ± 0%
Combine/slice_100_multi_error-8      193ns ± 0%       461ns ± 1%         194ns ± 1%

name \ alloc/op                     master.txt   original_pr.txt  simplified_pr.txt
Combine/inline_1-8                   16.0B ± 0%        0.0B               0.0B
Combine/inline_2-8                   32.0B ± 0%        0.0B               0.0B
Combine/inline_3_no_error-8          48.0B ± 0%        0.0B               0.0B
Combine/inline_3_one_error-8         48.0B ± 0%        0.0B               0.0B
Combine/inline_3_multiple_errors-8   80.0B ± 0%       80.0B ± 0%         80.0B ± 0%
Combine/slice_100_no_errors-8        0.00B            0.00B              0.00B
Combine/slice_100_one_error-8        0.00B            0.00B              0.00B
Combine/slice_100_multi_error-8      64.0B ± 0%     1856.0B ± 0%         64.0B ± 0%

name \ allocs/op                    master.txt   original_pr.txt  simplified_pr.txt
Combine/inline_1-8                    1.00 ± 0%        0.00               0.00
Combine/inline_2-8                    1.00 ± 0%        0.00               0.00
Combine/inline_3_no_error-8           1.00 ± 0%        0.00               0.00
Combine/inline_3_one_error-8          1.00 ± 0%        0.00               0.00
Combine/inline_3_multiple_errors-8    2.00 ± 0%        2.00 ± 0%          2.00 ± 0%
Combine/slice_100_no_errors-8         0.00             0.00               0.00
Combine/slice_100_one_error-8         0.00             0.00               0.00
Combine/slice_100_multi_error-8       2.00 ± 0%        3.00 ± 0%          2.00 ± 0%
```

The simplified version performs a nanosecond slower, does the same number of allocations or fewer in the case of "slice of 100 with multiple errors" (because it optimizes for the "no nested multierrors case").